Improve tests.
